Conflicts:
	CamShop/CamShop.csproj.user
	CamShop/Views/Cart/Index.cshtml
	CamShop/Views/SanPham/Detail.cshtml
	CamShop/bin/CamShop.dll
	CamShop/bin/CamShop.pdb
	CamShop/bin/Models.dll
	CamShop/bin/Models.pdb
	CamShop/obj/Debug/CamShop.csproj.CoreCompileInputs.cache
	CamShop/obj/Debug/CamShop.csprojAssemblyReference.cache
	CamShop/obj/Debug/CamShop.dll
	CamShop/obj/Debug/CamShop.pdb
	CamShop/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	Models/bin/Debug/Models.dll
	Models/bin/Debug/Models.pdb
	Models/obj/Debug/Models.dll
	Models/obj/Debug/Models.pdb

+++ Models/EF/ChiTietHoaDon.cs
@@ -18,7 +18,7 @@ namespace Models.EF
 
         public short? soLuong { get; set; }
 
-        public int? thanhTien { get; set; }
+        public double? thanhTien { get; set; }
 
         public virtual HoaDon HoaDon { get; set; }
